Typing API route, adding sample logic for the sankey graph
