+++ src/app/api/types.ts
@@ -1,0 +1,7 @@
+export interface PageExitEvent {
+  event_id: string;
+  page: string;
+  url: string;
+  session_id: string;
+  timestamp: number;
+}
